# scaffold/hr-service/openapi.yaml
openapi: 3.0.3
info:
  title: PMP Human Resources Service
  version: "1.0"
  description: >
    Implements FR-HR-01..07 (docs Section 4.6). FastAPI must generate a schema
    that matches this contract — treat mismatches as a bug, not a spec update,
    unless the SRS itself is being revised.

    All endpoints require a bearer access token (RS256 JWT) issued by
    iam-service; its signature is verified against iam-service's JWKS and the
    required `<resource>.<action>` permission must be present in the token's
    `permissions` claim (CLAUDE.md rule 4). Missing/invalid token -> 401;
    valid token without the permission -> 403.

    discipline_records are additionally restricted: only hr.discipline.read /
    hr.discipline.write unlock those routes (docs Section 2.3 — HR/command
    only). FR-HR-08 (roster/headcount reports) is not implemented — it reads
    to derive reports rather than defining new domain state, and is deferred.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /units:
    post:
      summary: Create an organisational unit
      description: Requires `hr.unit.write`. Publishes a `UnitCreated` domain event.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UnitCreate" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Unit" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.unit.write }
    get:
      summary: List organisational units
      description: Requires `hr.unit.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: station_id, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Units
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Unit" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.unit.read }
...
  /officers/{officer_id}/transfers:
    post:
      summary: Request a transfer (FR-HR-03)
      description: >
        Requires `hr.transfer.write`. Opens a `pending` transfer request;
        `from_unit_id` is the officer's current unit. Publishes a
        `TransferRequested` domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferCreate" }
      responses:
        "201": { description: Requested, content: { application/json: { schema: { $ref: "#/components/schemas/Transfer" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.transfer.write }
        "404": { description: No officer with that id, or to_unit_id does not exist }
    get:
      summary: Transfer history for one officer (FR-HR-03)
      description: Requires `hr.transfer.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Transfers, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Transfer" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.transfer.read }
        "404": { description: No officer with that id }
  /transfers:
    get:
      summary: Approver queue — every transfer request (FR-HR-03)
      description: Requires `hr.transfer.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [pending, approved, rejected] } }
      responses:
        "200":
          description: Transfers, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Transfer" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.transfer.read }
  /transfers/{transfer_id}:
    patch:
      summary: Approve or reject a pending transfer (FR-HR-03)
      description: >
        Requires `hr.transfer.approve`. On approval, updates the officer's
        current unit and closes/opens the corresponding assignment record
        (same effect as POST .../assignments). Publishes a
        `TransferStatusChanged` domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: transfer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferStatusUpdate" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Transfer" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.transfer.approve }
        "404": { description: No transfer with that id }
        "409": { description: Transfer is not pending }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token issued by iam-service (/api/v1/auth/mfa/verify).
  schemas:
    UnitCreate:
      type: object
      required: [name, station_id]
      properties:
        name: { type: string }
        station_id: { type: string, format: uuid }
    Unit:
      allOf:
        - $ref: "#/components/schemas/UnitCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
...
    TransferCreate:
      type: object
      required: [to_unit_id]
      properties:
        to_unit_id: { type: string, format: uuid }
    Transfer:
      allOf:
        - $ref: "#/components/schemas/TransferCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            officer_id: { type: string, format: uuid }
            from_unit_id: { type: string, format: uuid, nullable: true }
            status: { type: string, enum: [pending, approved, rejected] }
            created_at: { type: string, format: date-time }
    TransferStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, rejected] }
...